requirements.txt
discord.py==1.4.1
google
psycopg2-binary